RAG System/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
sentence-transformers
streamlit
python-dotenv
tiktoken
pandas
numpy
pytest
